Add like, comment, copy link buttons Add copy link functionality
