+++ src/data/miscFuncs.js
@@ -1,0 +1,9 @@
+const READING_SPEED = 100; // wpm
+const measureReadingTime = (str) => {
+  const nWords = str.trim().split(/\s+/).length;
+  return parseInt(nWords / READING_SPEED);
+};
+
+module.exports = {
+  measureReadingTime,
+};
